Autofill texts
Now with the scripts Bindings.js and Query.js it's easier to make the webpages. You just have to configure the route of the JSON and this scripts will repalce the text of the HTML by the text they received via JSON.

Also added exchange as an static example that can potentially implement this two scripts.